Add category metrics views and shared analytics module
